+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 
 ### Added
 
+* `--config` option to the CLI. It allows you to specify a configuration file other than `cucumber.js`.
+  See [docs/profiles.md](./docs/profiles.md#using-another-file-than-cucumberjs) for more info.
+  [#1794](https://github.com/cucumber/cucumber-js/pull/1794)
+
 ### Changed
 
 ### Deprecated